further implementation of DATA API usage

+++ include/core/np_comp_msgproperty.h
@@ -276,8 +276,8 @@ struct np_msgproperty_s
     // cache which will hold up to max_threshold messages
     np_msgcache_policy_type cache_policy;
     uint16_t  cache_size;
-    uint8_t  msg_threshold; // current cache size
-    uint8_t  max_threshold; // local cache size
+    uint32_t  msg_threshold; // current cache size
+    uint32_t  max_threshold; // local cache size
 
     bool is_internal;
